requirements.txt
numpy
pillow
streamlit
tensorflow
matplotlib
pandas
keras